job: Require users to setup and cleanup Job temporary handlers
Each Job creates several dirs/files and uses certain handlers to setup
the environment, let's be explicit and instead of preparing everything
magically on __init__ and cleaning most of it after "run" let users to
"setup" and "cleanup" the Job handlers explicitly and to simplify the
usage also support __enter__ and __exit__ methods to be able to simply
use:

    # No job
    with Job(...) as job:
        # do your stuff with temporary handlers set
    # Job temporary handlers cleaned

+++ optional_plugins/runner_remote/tests/test_remote.py
@@ -64,8 +64,10 @@ class RemoteTestRunnerTest(unittest.TestCase):
                                                  '/tests/other/test',
                                                  'passtest.py'])
 
+        job = None
         try:
             job = Job(job_args)
+            job.setup()
             runner = avocado_runner_remote.RemoteTestRunner(job, job.result)
             return_value = (True, (version.MAJOR, version.MINOR))
             runner.check_remote_avocado = mock.Mock(return_value=return_value)
@@ -102,7 +104,8 @@ class RemoteTestRunnerTest(unittest.TestCase):
                                                  ignore_status=True,
                                                  timeout=61)
         finally:
-            shutil.rmtree(job.args.base_logdir)
+            if job:
+                shutil.rmtree(job.args.base_logdir)
 
 
 if __name__ == '__main__':
